FaceRec/requirements.txt
numpy==1.26.4
opencv_python==4.8.0.76
Pillow==10.0.1
PyQt5==5.15.10
pyqt5_sip==12.18.0
openpyxl==3.1.5
dlib==19.24.6
